+++ src/GameDataStructs.hpp
@@ -6,8 +6,8 @@
 
 struct EntityDef {
   std::string name = "New Entity";
-  int width = 16;
-  int height = 16;
+  float width = 16;
+  float height = 16;
   std::string color = "white";
   std::string script = "myscript.lua";
 };
